feat: calculate quiz score

+++ backend/src/main/java/org/iamflubin/quiz/domain/Score.java
@@ -1,6 +1,6 @@
 package org.iamflubin.quiz.domain;
 
-public record Score(int value) {
+public record Score(double value) {
     public Score {
         if (value < 0) {
             throw new IllegalArgumentException("Score cannot be negative");
